+++ gatsby-config.js
@@ -11,6 +11,10 @@ module.exports = {
     `gatsby-plugin-typescript`,
     `gatsby-plugin-styled-components`,
     `gatsby-plugin-tslint`,
+    `gatsby-transformer-sharp`,
+    `gatsby-plugin-sharp`,
+    `gatsby-source-meetup-members`,
+    `gatsby-source-meetup-group`,
     {
       resolve: `gatsby-source-filesystem`,
       options: {
@@ -26,8 +30,6 @@ module.exports = {
         },
       },
     },
-    `gatsby-transformer-sharp`,
-    `gatsby-plugin-sharp`,
     {
       resolve: `gatsby-plugin-manifest`,
       options: {
@@ -40,7 +42,6 @@ module.exports = {
         icon: `src/images/lijs-icon.png`, // This path is relative to the root of the site.
       },
     },
-
     // this (optional) plugin enables Progressive Web App + Offline functionality
     // To learn more, visit: https://gatsby.dev/offline
     // `gatsby-plugin-offline`,