+++ src/components/animations/fade-in.tsx
@@ -6,15 +6,16 @@ const fadeInAnimation = keyframes`${fadeIn}`;
 
 interface FadeInProps {
   block?: boolean;
+  duration?: string;
 }
 
 const StyledFadeIn = styled.span<FadeInProps>`
-  display: ${(props) => (props.block ? 'block' : 'inline')}
-  animation: 1s ${fadeInAnimation};
+  display: ${(props) => (props.block ? 'block' : 'inline')};
+  animation: ${(props) => props.duration || '1'}s ${fadeInAnimation};
 `;
 
 const FadeIn: React.FC<FadeInProps> = ({ children, block }) => (
-  <StyledFadeIn block={true}>{children}</StyledFadeIn>
+  <StyledFadeIn block={block}>{children}</StyledFadeIn>
 );
 
 export default FadeIn;