+++ src/components/common/section.tsx
@@ -10,7 +10,7 @@ const Section = styled.section<SectionProps>`
   width: 100%;
   background-color: ${(props) => props.bg || 'white'};
   margin: 0 auto;
-  padding: 16px;
+  padding: 20px;
   padding-top: 20px;
   color: ${(props) => props.color || '#333'};
 `;